refactor: improve semantics of the comment time (#10)
Replace custom elements and styles built to show the full time on hover to
use a standard `time` element. This improves out-of-the-box support for
showing the full time on hover and improves the semantics of the HTML.
Add a new function to format a given time using an ISO format to be more
machine-readable. This format of time is used in the `datetime` attribute
of the new `time` element.
Refactor the existing code to format time in a human-readable format to be
hardcoded to UTC and use the `Sun Mar 14 16:23:15 2021 UTC` format.